Wine wordcloud plot

The wordcloud plot extracts key words from the description variable to visualize flavors and taste profiles as written by the reviewer.

Usage

wordcloud(input_country)

Arguments

input_country

A character of country of wine origin.

Value

A wordcloud2 object. The plot of wine descriptions. If the input_country is in the dataset, a wordcloud plot of wine descriptions for that country is generated. If the input_country is not in the dataset, or if an incorrect country is input, the wordcloud plot will be generated with all wine descriptions in the dataset.

Examples

wordcloud("US")
#> Selecting by n

wordcloud("world")
#> Selecting by n